+++ src/kekkonen/core.clj
@@ -59,6 +59,43 @@
    s/Keyword s/Any})
 
 ;;
+;; Interceptors
+;;
+
+(s/defschema Interceptor
+  (s/constrained
+    {(s/optional-key :name) s/Str
+     (s/optional-key :enter) Function
+     (s/optional-key :leave) Function}
+    (fn [{:keys [enter leave]}] (or enter leave))
+    'enter-or-leave-required))
+
+(s/defschema FunctionOrInterceptor
+  (s/conditional fn? Function :else Interceptor))
+
+(defn interceptor [interceptor-or-a-function]
+  (s/validate
+    Interceptor
+    (cond
+      (fn? interceptor-or-a-function) {:enter interceptor-or-a-function}
+      (map? interceptor-or-a-function) interceptor-or-a-function
+      :else (throw (ex-info (str "Can't coerce into an interceptor: " interceptor-or-a-function) {})))))
+
+(defn- interceptor-factory [data]
+  (assert (vector? data) "intercetors must be defined as a vector")
+  (let [interceptors (map (fn [x] (interceptor (if (vector? x) (apply (first x) (rest x)) x))) data)
+        execute (fn [[first & rest] ctx]
+                  (if-let [ctx (first ctx)]
+                    (if rest
+                      (recur rest ctx)
+                      ctx)))
+        enters (seq (keep :enter interceptors))
+        leaves (seq (reverse (keep :leave interceptors)))]
+    (merge
+      (if enters {:enter (partial execute enters)})
+      (if leaves {:leave (partial execute leaves)}))))
+
+;;
 ;; Type Resolution
 ;;
 
@@ -219,7 +256,7 @@
    context :- KeywordMap
    coercion :- {:input (s/maybe KeywordMap)
                 :output s/Any}
-   transformers :- [Function]
+   interceptors :- [Interceptor]
    user :- KeywordMap])
 
 (defmethod clojure.core/print-method Dispatcher
@@ -286,7 +323,7 @@
 
 (defn input-coerce!
   ([context schema]
-    ;; TODO: ensure that dispatcher is always present, also for transformers.
+    ;; TODO: ensure that dispatcher is always present, also for interceptors.
    (if-let [dispatcher (get-dispatcher context)]
      (input-coerce! context schema (-> dispatcher :coercion :input))
      (throw (ex-info "no attached dispatcher." {}))))
@@ -315,34 +352,41 @@
   [dispatcher, action :- s/Keyword]
   (get-in dispatcher [:handlers action]))
 
+(defn- invalid-action! [action]
+  (throw (ex-info (str "Invalid action") {:type ::dispatch, :value action})))
+
 (defn- dispatch [dispatcher mode action context]
   (if-let [{:keys [function all-user input output] :as handler} (some-handler dispatcher action)]
     (let [input-matcher (-> dispatcher :coercion :input)
           context (as-> context context
 
-                        ; TODO: transformers -> middleware/interceptors
-
                         ;; base-context from Dispatcher
                         (kc/deep-merge (:context dispatcher) context)
 
-                        ;; run all the transformers
-                        ;; short-circuit execution if a transformer returns nil
-                        (reduce (fn [ctx mapper] (or (mapper ctx) (reduced nil))) context (:transformers dispatcher))
+                        ;; run all the interceptor enters, short-circuit on nil
+                        (reduce
+                          (fn [ctx {:keys [enter]}]
+                            (if enter (or (enter ctx) (reduced nil)) ctx))
+                          context
+                          (:interceptors dispatcher))
 
-                        ; TODO: type-transformers?
+                        ; TODO: type-interceptors?
 
-                        ;; run all the user transformers per namespace/handler
+                        ;; run all the user interceptor enters per namespace/handler
                         ;; start from the root. a returned nil context short-circuits
                         ;; the run an causes ::dispatch error. Apply local coercion
                         ;; in the input is defined (using same definitions as with handlers)
                         ; TODO: precompile for fail-fast & speed? input-coerce only tagged ones? test!
                         (reduce
                           (fn [ctx [k v]]
-                            (if-let [mapper-gen (get-in dispatcher [:user k])]
-                              (let [mapper (mapper-gen v)
-                                    input-schema (:input (kc/extract-schema mapper))
-                                    ctx (input-coerce! ctx input-schema input-matcher)]
-                                (or (mapper ctx) (reduced nil)))
+                            (if-let [interceptor-factory (get-in dispatcher [:user k])]
+                              (if-let [interceptor (interceptor (interceptor-factory v))]
+                                (if-let [enter (:enter interceptor)]
+                                  (let [input-schema (:input (kc/extract-schema enter))
+                                        ctx (input-coerce! ctx input-schema input-matcher)]
+                                    (or (enter ctx) (reduced nil)))
+                                  ctx)
+                                ctx)
                               ctx))
                           context
                           (apply concat all-user))
@@ -356,16 +400,42 @@
                                                                 ::handler handler})))]
 
       (when-not context
-        (throw (ex-info (str "Invalid action") {:type ::dispatch, :value action})))
+        (invalid-action! action))
 
       ;; all good, let's invoke?
       (if (#{:invoke} mode)
-        (let [response (function context)]
-          ;; response coercion
-          (if (and output (-> dispatcher :coercion :output))
-            (coerce! output (-> dispatcher :coercion :output) response nil ::response)
-            response))))
-    (throw (ex-info (str "Invalid action") {:type ::dispatch, :value action}))))
+        (let [response (as-> (function context) response
+                             (if (and output (-> dispatcher :coercion :output))
+                               (coerce! output (-> dispatcher :coercion :output) response nil ::response)
+                               response))
+              context (as-> (assoc context :response response) context
+
+                            ;; run all the user interceptor leaves per namespace/handler
+                            ;; in reverse order start from the handler. a returned nil context short-circuits
+                            ;; the run an causes ::dispatch error. Apply local coercion
+                            ;; in the input is defined (using same definitions as with handlers)
+                            (reduce
+                              (fn [ctx [k v]]
+                                (if-let [interceptor-factory (get-in dispatcher [:user k])]
+                                  (if-let [interceptor (interceptor (interceptor-factory v))]
+                                    (if-let [leave (:leave interceptor)]
+                                      (or (leave ctx) (reduced nil))
+                                      ctx)
+                                    ctx)
+                                  ctx))
+                              context
+                              (reverse (apply concat all-user)))
+
+                            ;; run all the interceptor leaves in reverse order, short-circuit on nil
+                            (reduce
+                              (fn [ctx {:keys [leave]}]
+                                (if leave (or (leave ctx) (reduced nil)) ctx))
+                              context
+                              (reverse (:interceptors dispatcher))))]
+
+          (or (:response context)
+              (invalid-action! action)))))
+    (invalid-action! action)))
 
 (s/defn check
   "Checks an action handler with the given context."
@@ -454,7 +524,7 @@
   {:handlers {(s/cond-pre s/Keyword Namespace) s/Any}
    (s/optional-key :context) KeywordMap
    (s/optional-key :type-resolver) Function
-   (s/optional-key :transformers) [Function]
+   (s/optional-key :interceptors) [FunctionOrInterceptor]
    (s/optional-key :coercion) {(s/optional-key :input) (s/maybe KeywordMap)
                                (s/optional-key :output) s/Any}
    (s/optional-key :user) (s/cond-pre [[(s/one s/Keyword 'key) Function]] KeywordMap)
@@ -463,12 +533,13 @@
 (s/def +default-options+ :- Options
   {:handlers {}
    :context {}
-   :transformers []
+   :interceptors []
    :coercion {:input {:data (constantly nil)}
               :output (constantly nil)}
    :type-resolver default-type-resolver
-   :user {}})
+   :user {:interceptors interceptor-factory}})
 
+;; TODO: create full set of interceptors here and run them in order
 (defn- collect-and-enrich
   [{:keys [handlers type-resolver user]} allow-empty-namespaces?]
   (let [handler-ns (fn [m] (if (seq m) (->> m (map :name) (map name) (str/join ".") keyword)))
@@ -491,7 +562,7 @@
                          all-user (map reorder (if-not (empty? user-meta) (conj ns-user user-meta) ns-user))
                          user-input (reduce
                                       (fn [acc [k v]]
-                                        (if-let [f (user k)]
+                                        (if-let [f (:enter (interceptor (user k)))]
                                           (let [schema (:input (kc/extract-schema (f v)))]
                                             (kc/merge-map-schemas acc schema))
                                           acc)) {} (apply concat all-user))
@@ -522,16 +593,15 @@
   "Creates a Dispatcher"
   [options :- Options]
   (let [options (-> options
-                    (->> (kc/deep-merge +default-options+))
-                    (update :user (fn [user]
-                                    (if-not (map? user)
-                                      (apply linked/map (apply concat user))
-                                      user))))
-        handlers (->> (collect-and-enrich options false))]
+                    (update :user (partial into (linked/map)))
+                    (->> (kc/deep-merge +default-options+)))
+        handlers (->> (collect-and-enrich options false))
+        interceptors (mapv interceptor (:interceptors options))]
     (map->Dispatcher
       (merge
-        (select-keys options [:context :transformers :coercion :user])
-        {:handlers handlers}))))
+        (select-keys options [:context :coercion :user])
+        {:handlers handlers
+         :interceptors interceptors}))))
 
 (s/defn transform-handlers
   "Applies f to all handlers. If the call returns nil,